+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#33](https://github.com/LaxarJS/grunt-init-laxar-application/issues/33): example widgets: added missing spec tests
+
 
 ## v1.0.3